Update board-staff.md
This is the text that was removed from the Staff Bios...keeping this as a record:

## Mason C. Clutter, Attorney-Advisor
**Email: mclutter@pclob.gov**

Mason Clutter joined the Privacy and Civil Liberties Oversight Board in September 2014 as an Attorney-Advisor to the Board. Ms. Clutter previously served as National Security and Privacy Counsel to the National Association of Criminal Defense Lawyers, where she worked on issues ranging from the detention and trial of terrorism suspects to government surveillance and the Fourth Amendment in the digital age. She also served as the Association’s official observer to the military commissions proceedings at Guantanamo Bay, Cuba. Before that, Ms. Clutter served as Counsel to the Rule of Law Program at the Constitution Project and as a Research Associate with the Lowenstein International Human Rights Law Clinic at Yale Law School. She began her legal career as an associate litigator at a civil defense firm in Orlando, Florida. Ms. Clutter holds an LL.M degree in Law and Government with a specialization in civil and constitutional rights from the American University Washington College of Law and a J.D. degree from Barry University School of Law. She is a graduate of the University of Central Florida and holds a B.A. degree in Legal Studies.

## Lynn Parker Dupree, Deputy General Counsel
Lynn Parker Dupree joined the Privacy and Civil Liberties Oversight Board in April 2014 and currently serves as the Deputy General Counsel. Prior to joining the PCLOB, Mrs. Dupree was detailed to the White House as a Deputy Associate Counsel for Presidential Personnel. She previously worked as the Special Assistant to the General Counsel at the Department of Homeland Security and also served as the acting Office of the General Counsel Executive Secretariat. She worked on a varied portfolio of issues, particularly issues related to privacy, information sharing, civil rights, and civil liberties. Prior to joining the Office of the General Counsel, Mrs. Dupree was the Special Assistant to the Chief Privacy Officer at DHS. In that role, her work focused on the Department of Homeland Security's participation in the Information Sharing Environment. Immediately prior to her appointment to the Department of Homeland Security Mrs. Dupree was an associate in the litigation practice group of Burr Forman LLP. She was a law clerk to Chief Judge Michael P. Mills in the Northern District of Mississippi. Mrs. Dupree graduated summa cum laude from Mississippi State University with a degree in Business Administration and received her law degree from the University of Mississippi.

## Prem M. Trivedi, Attorney-Advisor
**Email: ptrivedi@pclob.gov**

Prem Trivedi is an Attorney-Advisor to the Privacy and Civil Liberties Oversight Board. He joined the PCLOB in 2014 as Counselor to former Board Member James Dempsey. He previously served as a Legal Fellow with the Public International Law & Policy Group and as an associate in the International Arbitration group at Shearman & Sterling LLP. Mr. Trivedi is a lecturer at Georgetown University, where he teaches a seminar on U.S. foreign policy in South Asia. He publishes on international and domestic policy issues in South Asia, and has worked in India as a Fulbright Scholar, with a research institute, and with a nonprofit organization focused on governance reform. Mr. Trivedi is a recipient of Harvard Law School’s Heyman Fellowship. He is a graduate of Harvard Law School and Columbia University.